# apps/docs-site/static/openapi.yaml
openapi: 3.1.0
info:
  title: Spiritual Q&A Backend
  version: 0.1.0
paths:
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      description: Register a new user with E2EE key material.
      operationId: register_api_auth_register_post
      parameters:
      - name: use_cookies
        in: query
        required: false
        schema:
          type: boolean
          description: Return tokens in HttpOnly cookies
          default: false
          title: Use Cookies
        description: Return tokens in HttpOnly cookies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Login Challenge
      description: 'Step 1: Get user''s salt for client-side key derivation.'
      operationId: login_challenge_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginChallengeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginChallengeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login/verify:
    post:
      tags:
      - Authentication
      summary: Login Verify
      description: 'Step 2: Verify client auth token and return JWT pair + wrapped
        AK.'
      operationId: login_verify_api_auth_login_verify_post
      parameters:
      - name: use_cookies
        in: query
        required: false
        schema:
          type: boolean
          description: Return tokens in HttpOnly cookies
          default: false
          title: Use Cookies
        description: Return tokens in HttpOnly cookies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginVerifyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh Tokens
      description: Rotate refresh token and issue new JWT pair.
      operationId: refresh_tokens_api_auth_refresh_post
      parameters:
      - name: use_cookies
        in: query
        required: false
        schema:
          type: boolean
          description: Return tokens in HttpOnly cookies
          default: false
          title: Use Cookies
        description: Return tokens in HttpOnly cookies
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/RefreshRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/logout:
    post:
      tags:
      - Authentication
      summary: Logout
      description: Revoke the current refresh token.
      operationId: logout_api_auth_logout_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__schemas__common_schemas__MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/account:
    delete:
      tags:
      - Authentication
      summary: Delete Account
      description: Permanently delete user account and all data.
      operationId: delete_account_api_auth_account_delete
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__schemas__common_schemas__MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/change-password:
    post:
      tags:
      - Authentication
      summary: Change Password
      description: Update auth credentials after client-side password change.
      operationId: change_password_api_auth_change_password_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__schemas__common_schemas__MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/recover:
    post:
      tags:
      - Authentication
      summary: Recover Account
      description: Recover account with new credentials (after mnemonic verification).
      operationId: recover_account_api_auth_recover_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverAccountRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoverAccountResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuthResponse:
      properties:
        user_id:
          type: string
          title: User Id
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        access_expires_at:
          type: string
          format: date-time
          title: Access Expires At
      type: object
      required:
      - user_id
      - access_token
      - refresh_token
      - access_expires_at
      title: AuthResponse
      description: Response for registration (no wrapped AK return needed).
...
    ChangePasswordRequest:
      properties:
        new_auth_token:
          type: string
          minLength: 1
          title: New Auth Token
        new_wrapped_account_key:
          type: string
          minLength: 1
          title: New Wrapped Account Key
      type: object
      required:
      - new_auth_token
      - new_wrapped_account_key
      title: ChangePasswordRequest
      description: 'E2EE password change: client re-wraps AK with new LMK.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginChallengeRequest:
      properties:
        email:
          type: string
          maxLength: 255
          format: email
          title: Email
      type: object
      required:
      - email
      title: LoginChallengeRequest
      description: 'Step 1 of E2EE login: client sends email to get salt.'
    LoginChallengeResponse:
      properties:
        salt:
          type: string
          title: Salt
          description: Base64-encoded Argon2id salt
        recovery_wrapped_ak:
          anyOf:
          - type: string
          - type: 'null'
          title: Recovery Wrapped Ak
          description: Base64-encoded RecoveryKey-wrapped AccountKey (used for recovery)
      type: object
      required:
      - salt
      title: LoginChallengeResponse
      description: Server returns the user's salt for key derivation.
    LoginVerifyRequest:
      properties:
        email:
          type: string
          maxLength: 255
          format: email
          title: Email
        client_auth_token:
          type: string
          minLength: 1
          title: Client Auth Token
          description: Base64-encoded HKDF-derived auth token from LMK
      type: object
      required:
      - email
      - client_auth_token
      title: LoginVerifyRequest
      description: 'Step 2: client sends derived auth token for verification.'
    LoginVerifyResponse:
      properties:
        user_id:
          type: string
          title: User Id
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        access_expires_at:
          type: string
          format: date-time
          title: Access Expires At
          description: ISO 8601 datetime when the access token expires
        wrapped_account_key:
          type: string
          title: Wrapped Account Key
          description: Client uses this to unwrap AK with LMK
      type: object
      required:
      - user_id
      - access_token
      - refresh_token
      - access_expires_at
      - wrapped_account_key
      title: LoginVerifyResponse
      description: Successful login returns JWT pair + wrapped AK.
...
    RecoverAccountRequest:
      properties:
        email:
          type: string
          maxLength: 255
          format: email
          title: Email
        new_auth_token:
          type: string
          minLength: 1
          title: New Auth Token
        new_wrapped_account_key:
          type: string
          minLength: 1
          title: New Wrapped Account Key
      type: object
      required:
      - email
      - new_auth_token
      - new_wrapped_account_key
      title: RecoverAccountRequest
      description: 'E2EE account recovery: client unwraps AK with RK, re-wraps with
        new LMK.'
    RecoverAccountResponse:
      properties:
        user_id:
          type: string
          title: User Id
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        access_expires_at:
          type: string
          format: date-time
          title: Access Expires At
        wrapped_account_key:
          type: string
          title: Wrapped Account Key
      type: object
      required:
      - user_id
      - access_token
      - refresh_token
      - access_expires_at
      - wrapped_account_key
      title: RecoverAccountResponse
      description: Successful recovery auto-logs in and returns new tokens.
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          minLength: 1
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshRequest
      description: Refresh token request (mobile sends in body, web via cookie).
    RefreshResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        access_expires_at:
          type: string
          format: date-time
          title: Access Expires At
      type: object
      required:
      - access_token
      - refresh_token
      - access_expires_at
      title: RefreshResponse
      description: New token pair after refresh.
    RegisterRequest:
      properties:
        email:
          type: string
          maxLength: 255
          format: email
          title: Email
        client_auth_token:
          type: string
          minLength: 1
          title: Client Auth Token
          description: Base64-encoded HKDF-derived auth token from LMK
        salt:
          type: string
          minLength: 1
          title: Salt
          description: Base64-encoded Argon2id salt (16 bytes)
        wrapped_account_key:
          type: string
          minLength: 1
          title: Wrapped Account Key
          description: Base64-encoded AES-GCM wrapped AccountKey
        recovery_wrapped_ak:
          type: string
          minLength: 1
          title: Recovery Wrapped Ak
          description: Base64-encoded RecoveryKey-wrapped AccountKey
      type: object
      required:
      - email
      - client_auth_token
      - salt
      - wrapped_account_key
      - recovery_wrapped_ak
      title: RegisterRequest
      description: 'E2EE registration: client sends derived auth token + wrapped keys.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    app__schemas__common_schemas__MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
      description: Simple message response for operations like logout.